# app2/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Metal API",
    "description": "API for managing metal data",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/metals": {
      "get": {
        "summary": "Get all metals",
        "responses": {
          "200": {
            "description": "List of metals"
          }
        }
      },
      "post": {
        "summary": "Add a new metal",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Metal object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "price_per_kg": { "type": "number" },
                "density": { "type": "number" },
                "category": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Metal added"
          }
        }
      }
    },
    "/metals/{id}": {
      "get": {
        "summary": "Get metal by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Metal found" },
          "404": { "description": "Metal not found" }
        }
      },
      "put": {
        "summary": "Update metal by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "price_per_kg": { "type": "number" },
                "density": { "type": "number" },
                "category": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Updated" }
        }
      },
      "delete": {
        "summary": "Delete metal by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}